In TrinaryTreeViewController, change selector nodeButtonTapped to nodeButtonTapped: that passes it's sender. Each nodeButton calls the same selector nodeButtonTapped:. Each nodeButton is a standard UIButton, which is a subclass of UIView. UIView has a settable tag property. In this commit, set the nodeButton tag property to identify the nodeButton.
Add property nodeButtonDictionary to associate each tag as a key to a node as a value.
nodeButtonTapped: uses the sender's tag as a key to get the node from the nodeButtonDictionary.

Can't use nodeContent value as a tag, because it's not a unique key. The tree can contain multiple nodes with the same value.
Instead add instance variable buttonTagIndex and increment it for each button tag.

This approach allows the MVC View to use a standard UIButton without any additional knowledge or properties of the Model such as the node.
The MVC Controller TrinaryTreeViewController manages the dictionary nodeButtonDictionary to coordinate between the View and the Model.

This seems more consistent with Cocoa MVC than previous approaches I tried, such as subclassing UIButton or making a NodeButton object composited from a node (Model) and a button (View) property.
Reference http://stackoverflow.com/questions/450222/passing-a-parameter-in-setaction